third_party/blink/web_tests/external/wpt/html/webappapis/dynamic-markup-insertion/document-write/contentType.window-expected.txt

This is a testharness.js-based test.
FAIL document.write(): video document assert_equals: expected "application/ogg" but got "video/ogg"
FAIL document.write(): image document Cannot read properties of null (reading 'firstChild')
PASS document.write(): text document
PASS document.write(): HTML document
Harness: the test ran to completion.
